FastQC-viz: Creat all help

When creating the replacement list for fill the HTML template file, use this to add help sections. It creates keys for "help_" to each module.

Usage

add_all_help(add_help)

Value

list, list with all the "help_" keys for plots and headers.

Examples

add_all_help(add_help = TRUE)
#> $help_per_sequence_quality_scores
#> <div class="card-face card-back" id="help-per_sequence_quality_scores">
#>   <p>Allows you to see if a subset of your sequences have universally low quality values.</p>
#>   <table>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #afe0b7ff;
#>   border: 1px solid #2f6638ff;
#>   color: #2f6638ff;
#> ">
#>   PASS
#> </span> </td>
#>    <td style="text-align:left;"> The most frequently observed mean quality is >= 27. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #e8d28fff;
#>   border: 1px solid #8d6c08ff;
#>   color: #8d6c08ff;
#> ">
#>   WARN
#> </span> </td>
#>    <td style="text-align:left;"> The most frequently observed mean quality is < 27. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #f0c6bcff;
#>   border: 1px solid #862a13ff;
#>   color: #862a13ff;
#> ">
#>   FAIL
#> </span> </td>
#>    <td style="text-align:left;"> The most frequently observed mean quality is < 20. </td>
#>   </tr>
#> </tbody>
#> </table>
#>   <p><a href="https://www.bioinformatics.babraham.ac.uk/projects/fastqc/Help/3%20Analysis%20Modules/3%20Per%20Sequence%20Quality%20Scores.html" target="_blank">See more ></a></p>
#> </div>
#> 
#> $help_per_base_sequence_quality
#> <div class="card-face card-back" id="help-per_base_sequence_quality">
#>   <p>Shows an overview of the range of quality values across all bases at each position in the FastQ file.</p>
#>   <table>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #afe0b7ff;
#>   border: 1px solid #2f6638ff;
#>   color: #2f6638ff;
#> ">
#>   PASS
#> </span> </td>
#>    <td style="text-align:left;"> Quality scores are within acceptable limits. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #e8d28fff;
#>   border: 1px solid #8d6c08ff;
#>   color: #8d6c08ff;
#> ">
#>   WARN
#> </span> </td>
#>    <td style="text-align:left;"> The lower quartile for any base is < 10, or the median is < 25. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #f0c6bcff;
#>   border: 1px solid #862a13ff;
#>   color: #862a13ff;
#> ">
#>   FAIL
#> </span> </td>
#>    <td style="text-align:left;"> The lower quartile for any base is < 5, or the median is < 20. </td>
#>   </tr>
#> </tbody>
#> </table>
#>   <p><a href="https://www.bioinformatics.babraham.ac.uk/projects/fastqc/Help/3%20Analysis%20Modules/2%20Per%20Base%20Sequence%20Quality.html" target="_blank">See more ></a></p>
#> </div>
#> 
#> $help_per_base_sequence_content
#> <div class="card-face card-back" id="help-per_base_sequence_content">
#>   <p>Plots out the proportion of each of the four normal DNA bases for each base position.</p>
#>   <table>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #afe0b7ff;
#>   border: 1px solid #2f6638ff;
#>   color: #2f6638ff;
#> ">
#>   PASS
#> </span> </td>
#>    <td style="text-align:left;"> The difference between A/T and G/C is <= 10% at all positions. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #e8d28fff;
#>   border: 1px solid #8d6c08ff;
#>   color: #8d6c08ff;
#> ">
#>   WARN
#> </span> </td>
#>    <td style="text-align:left;"> The difference between A/T or G/C is > 10% at any position. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #f0c6bcff;
#>   border: 1px solid #862a13ff;
#>   color: #862a13ff;
#> ">
#>   FAIL
#> </span> </td>
#>    <td style="text-align:left;"> The difference between A/T or G/C is > 20% at any position. </td>
#>   </tr>
#> </tbody>
#> </table>
#>   <p><a href="https://www.bioinformatics.babraham.ac.uk/projects/fastqc/Help/3%20Analysis%20Modules/4%20Per%20Base%20Sequence%20Content.html" target="_blank">See more ></a></p>
#> </div>
#> 
#> $help_per_base_n_content
#> <div class="card-face card-back" id="help-per_base_n_content">
#>   <p>Plots the percentage of base calls at each position for which an N was called.</p>
#>   <table>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #afe0b7ff;
#>   border: 1px solid #2f6638ff;
#>   color: #2f6638ff;
#> ">
#>   PASS
#> </span> </td>
#>    <td style="text-align:left;"> N content is <= 5% at all positions. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #e8d28fff;
#>   border: 1px solid #8d6c08ff;
#>   color: #8d6c08ff;
#> ">
#>   WARN
#> </span> </td>
#>    <td style="text-align:left;"> Any position shows an N content of > 5%. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #f0c6bcff;
#>   border: 1px solid #862a13ff;
#>   color: #862a13ff;
#> ">
#>   FAIL
#> </span> </td>
#>    <td style="text-align:left;"> Any position shows an N content of > 20%. </td>
#>   </tr>
#> </tbody>
#> </table>
#>   <p><a href="https://www.bioinformatics.babraham.ac.uk/projects/fastqc/Help/3%20Analysis%20Modules/6%20Per%20Base%20N%20Content.html" target="_blank">See more ></a></p>
#> </div>
#> 
#> $help_per_sequence_gc_content
#> <div class="card-face card-back" id="help-per_sequence_gc_content">
#>   <p>Measures the GC content across the whole length of each sequence and compares it to a modelled normal distribution.</p>
#>   <table>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #afe0b7ff;
#>   border: 1px solid #2f6638ff;
#>   color: #2f6638ff;
#> ">
#>   PASS
#> </span> </td>
#>    <td style="text-align:left;"> Deviations from the normal distribution are <= 15% of reads. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #e8d28fff;
#>   border: 1px solid #8d6c08ff;
#>   color: #8d6c08ff;
#> ">
#>   WARN
#> </span> </td>
#>    <td style="text-align:left;"> The sum of deviations from the normal distribution is > 15% of reads. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #f0c6bcff;
#>   border: 1px solid #862a13ff;
#>   color: #862a13ff;
#> ">
#>   FAIL
#> </span> </td>
#>    <td style="text-align:left;"> The sum of deviations from the normal distribution is > 30% of reads. </td>
#>   </tr>
#> </tbody>
#> </table>
#>   <p><a href="https://www.bioinformatics.babraham.ac.uk/projects/fastqc/Help/3%20Analysis%20Modules/5%20Per%20Sequence%20GC%20Content.html" target="_blank">See more ></a></p>
#> </div>
#> 
#> $help_sequence_length_distribution
#> <div class="card-face card-back" id="help-sequence_length_distribution">
#>   <p>Generates a graph showing the distribution of fragment sizes.</p>
#>   <table>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #afe0b7ff;
#>   border: 1px solid #2f6638ff;
#>   color: #2f6638ff;
#> ">
#>   PASS
#> </span> </td>
#>    <td style="text-align:left;"> All sequences have the same, non-zero length. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #e8d28fff;
#>   border: 1px solid #8d6c08ff;
#>   color: #8d6c08ff;
#> ">
#>   WARN
#> </span> </td>
#>    <td style="text-align:left;"> Sequences are not all the same length. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #f0c6bcff;
#>   border: 1px solid #862a13ff;
#>   color: #862a13ff;
#> ">
#>   FAIL
#> </span> </td>
#>    <td style="text-align:left;"> Any sequence has zero length. </td>
#>   </tr>
#> </tbody>
#> </table>
#>   <p><a href="https://www.bioinformatics.babraham.ac.uk/projects/fastqc/Help/3%20Analysis%20Modules/7%20Sequence%20Length%20Distribution.html" target="_blank">See more ></a></p>
#> </div>
#> 
#> $help_sequence_duplication_levels
#> <div class="card-face card-back" id="help-sequence_duplication_levels">
#>   <p>Counts the degree of duplication for every sequence in a library.</p>
#>   <table>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #afe0b7ff;
#>   border: 1px solid #2f6638ff;
#>   color: #2f6638ff;
#> ">
#>   PASS
#> </span> </td>
#>    <td style="text-align:left;"> Non-unique sequences are <= 20% of the total. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #e8d28fff;
#>   border: 1px solid #8d6c08ff;
#>   color: #8d6c08ff;
#> ">
#>   WARN
#> </span> </td>
#>    <td style="text-align:left;"> Non-unique sequences make up > 20% of the total. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #f0c6bcff;
#>   border: 1px solid #862a13ff;
#>   color: #862a13ff;
#> ">
#>   FAIL
#> </span> </td>
#>    <td style="text-align:left;"> Non-unique sequences make up > 50% of the total. </td>
#>   </tr>
#> </tbody>
#> </table>
#>   <p><a href="https://www.bioinformatics.babraham.ac.uk/projects/fastqc/Help/3%20Analysis%20Modules/8%20Duplicate%20Sequences.html" target="_blank">See more ></a></p>
#> </div>
#> 
#> $help_overrepresented_sequences
#> <div class="card-face card-back" id="help-overrepresented_sequences">
#>   <p>Lists all of the sequences which make up more than 0.1% of the total.</p>
#>   <table>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #afe0b7ff;
#>   border: 1px solid #2f6638ff;
#>   color: #2f6638ff;
#> ">
#>   PASS
#> </span> </td>
#>    <td style="text-align:left;"> No sequence represents > 0.1% of the total. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #e8d28fff;
#>   border: 1px solid #8d6c08ff;
#>   color: #8d6c08ff;
#> ">
#>   WARN
#> </span> </td>
#>    <td style="text-align:left;"> A sequence represents > 0.1% of the total. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #f0c6bcff;
#>   border: 1px solid #862a13ff;
#>   color: #862a13ff;
#> ">
#>   FAIL
#> </span> </td>
#>    <td style="text-align:left;"> A sequence represents > 1% of the total. </td>
#>   </tr>
#> </tbody>
#> </table>
#>   <p><a href="https://www.bioinformatics.babraham.ac.uk/projects/fastqc/Help/3%20Analysis%20Modules/9%20Overrepresented%20Sequences.html" target="_blank">See more ></a></p>
#> </div>
#> 
#> $help_adapter_content
#> <div class="card-face card-back" id="help-adapter_content">
#>   <p>Finds sequences (k-mers) which do not have even coverage, which can indicate adapter contamination.</p>
#>   <table>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #afe0b7ff;
#>   border: 1px solid #2f6638ff;
#>   color: #2f6638ff;
#> ">
#>   PASS
#> </span> </td>
#>    <td style="text-align:left;"> Adapter presence is <= 5% of all reads. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #e8d28fff;
#>   border: 1px solid #8d6c08ff;
#>   color: #8d6c08ff;
#> ">
#>   WARN
#> </span> </td>
#>    <td style="text-align:left;"> An adapter is present in > 5% of all reads. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #f0c6bcff;
#>   border: 1px solid #862a13ff;
#>   color: #862a13ff;
#> ">
#>   FAIL
#> </span> </td>
#>    <td style="text-align:left;"> An adapter is present in > 10% of all reads. </td>
#>   </tr>
#> </tbody>
#> </table>
#>   <p><a href="https://www.bioinformatics.babraham.ac.uk/projects/fastqc/Help/3%20Analysis%20Modules/10%20Adapter%20Content.html" target="_blank">See more ></a></p>
#> </div>
#>