FastQC-viz: Create Help

Create help sections for each module

Usage

create_help(module_name)

Arguments

module_name

character, name of the module

Value

character, The HTML elements that compose the header

Examples

create_help("per_base_sequence_quality")
#> <div class="card-face card-back" id="help-per_base_sequence_quality">
#>   <p>Shows an overview of the range of quality values across all bases at each position in the FastQ file.</p>
#>   <table>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #afe0b7ff;
#>   border: 1px solid #2f6638ff;
#>   color: #2f6638ff;
#> ">
#>   PASS
#> </span> </td>
#>    <td style="text-align:left;"> Quality scores are within acceptable limits. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #e8d28fff;
#>   border: 1px solid #8d6c08ff;
#>   color: #8d6c08ff;
#> ">
#>   WARN
#> </span> </td>
#>    <td style="text-align:left;"> The lower quartile for any base is < 10, or the median is < 25. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span class="pill" 
#> style="
#>   background: #f0c6bcff;
#>   border: 1px solid #862a13ff;
#>   color: #862a13ff;
#> ">
#>   FAIL
#> </span> </td>
#>    <td style="text-align:left;"> The lower quartile for any base is < 5, or the median is < 20. </td>
#>   </tr>
#> </tbody>
#> </table>
#>   <p><a href="https://www.bioinformatics.babraham.ac.uk/projects/fastqc/Help/3%20Analysis%20Modules/2%20Per%20Base%20Sequence%20Quality.html" target="_blank">See more ></a></p>
#> </div>