Format Numbers with Metric Suffixes (K, M, B, T)

This function takes a numeric vector and converts it into a character vector, formatting large numbers with appropriate suffixes for thousands (K), millions (M), billions (B), and trillions (T).

Usage

format_large_numbers(nums, digits = 0)

Arguments

nums

A numeric vector of numbers to format.

digits

An integer specifying the number of decimal places to use. Default is 1.

Value

A character vector of formatted numbers.

Examples

format_large_numbers(1234)
#> [1] "1 K"
#> [1] "1 K"

format_large_numbers(c(100, 5500, 1200000, -2500000000, 3450000000000))
#> [1] "100"  "6 K"  "1 M"  "-2 B" "3 T" 
#> [1] "100"    "5 K"  "1 M"  "-2 B" "3 T"

format_large_numbers(999999, digits = 2)
#> [1] "1000.00 K"
#> [1] "1.00 M"