FastQC-viz: Status to Icon

Create HTML header with the Status. This function is decrepted, it was used to add iconify in the report back when it was created with Quarto. Now it was replaced by status_to_pill().

Usage

status_to_icon(status, add_color = TRUE)

Arguments

status

character, "pass", "warn" or "fail"

add_color

boolean, add color

Value

status

Details

This function generates an HTML <span> element styled to look like a status pill or badge, similar to those used in the FastQC report interface.

The appearance of the pill is determined by the status argument:

"pass": Typically styled in green.

"warn": Typically styled in yellow.

"fail": Typically styled in red.

Examples

status_to_icon("pass")
#> [1] "<span style=\"color:#4ba359ff\">{{< iconify material-symbols:check-circle-rounded >}}</span>"